+++ apps/web/src/Constants/socket.constant.ts
@@ -1,0 +1,1 @@
+export const NEW_MESSAGE_RECEIVED_SOCKET_EVENT = 'new_message_received';
