+++ apps/web/src/Modules/teaminbox/components/chat.message.detail.component.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ChatMessageDetailComponent = () => {
+    return (
+        <div className='col-span-6 h-full rounded bg-base-100'>
+            ChatMessageDetailComponent
+        </div>
+    );
+};
+
+export default ChatMessageDetailComponent;
