+++ apps/web/src/Modules/teaminbox/components/chat.message.info.section.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ChatMessageInfoSection = () => {
+    return (
+        <div className='col-span-3 h-full rounded-lg bg-base-100'>
+            ChatMessageInfoSection
+        </div>
+    );
+};
+
+export default ChatMessageInfoSection;
