+++ packages/core/src/Constants/socketEvent.constant.ts
@@ -1,5 +1,3 @@
 export const WS_LOGIN = 'login';
-
 export const WS_ON_LOGIN = 'user_on_login';
-
 export const ON_SUCCESSFUL_CONNECTION = 'on_successful_connection';
